% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_estim_ncpMCA.R
\name{estim_ncpMCA_kfold}
\alias{estim_ncpMCA_kfold}
\title{Perform `modded_estim_ncpMCA` with K-fold CV}
\usage{
estim_ncpMCA_kfold(
  don,
  vrai.tab,
  ncp.min,
  ncp.max,
  nbsim,
  pNA,
  threshold,
  verbose
)
}
\value{
a list containing the estimated ncp and the corresponding criterion values.
}
\description{
Estimates the number of dimensions to retain in Multiple Correspondence Analysis (MCA) using K-fold cross-validation.
}
\details{
This function performs K-fold cross-validation to estimate the number of dimensions to retain in MCA.
It introduces missing values into the input data.frame using the `prodna1` function.
For each simulation and each number of dimensions (ncp), it computes the imputed data.frame using the `modded_imputeMCA` function.
The sum of squared differences between the imputed and true tables is calculated and stored in the res matrix.
The criterion values are computed by taking the mean of the sum of squared differences across simulations for each ncp.
The estimated number of dimensions (ncp) is determined by selecting the ncp value with the minimum criterion value.
}
