% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_estim_ncpMCA.R
\name{estim_ncpMCA_loo}
\alias{estim_ncpMCA_loo}
\title{Perform `modded_estim_ncpMCA` with LOO-CV}
\usage{
estim_ncpMCA_loo(don, ncp.min, ncp.max, threshold, verbose)
}
\value{
a list containing the estimated ncp and the corresponding criterion values.
}
\description{
Estimates the number of dimensions to retain in Multiple Correspondence Analysis (MCA) using LOO-CV.
}
\details{
This function performs leave-one-out cross-validation to estimate the number of dimensions to retain in MCA.
It iterates over each element of the input data.frame and introduces a missing value at that position.
For each missing value and each number of dimensions (ncp), it computes the imputed value using the `modded_imputeMCA` function.
The imputed values are stored in the tab.disj.hat matrix.
The criterion values are computed by taking the mean of the squared differences between the imputed and true tables.
The estimated number of dimensions (ncp) is determined by selecting the ncp value with the minimum criterion value.
}
