% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeMCA.R
\name{find.category.1}
\alias{find.category.1}
\title{Assign Category Based on Most Likely tabdisj}
\usage{
find.category.1(X, tabdisj)
}
\arguments{
\item{X}{original data.frame}

\item{tabdisj}{tabdisj object}
}
\value{
imputed data.frame with values based on most likely category
}
\description{
Assign Category Based on Most Likely tabdisj
}
