% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeMCA.R
\name{fit_compare_fns}
\alias{fit_compare_fns}
\title{Compare Results of imputeMCA and modded_imputeMCA}
\usage{
fit_compare_fns(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{arguments passed to imputeMCA and modded_imputeMCA}
}
\value{
test and modded_test
}
\description{
Compare Results of imputeMCA and modded_imputeMCA
}
