% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_estim_ncpPCA.R
\name{methyl_estim_ncpPCA}
\alias{methyl_estim_ncpPCA}
\title{Estimate ncp for Methyl Data}
\usage{
methyl_estim_ncpPCA(
  X,
  ncp.min = 0,
  ncp.max = 5,
  method = c("Regularized", "EM"),
  scale = TRUE,
  method.cv = c("gcv", "loo", "Kfold"),
  nbsim = 100,
  pNA = 0.05,
  threshold = 1e-04,
  verbose = TRUE,
  ...
)
}
\value{
list of value with and best ncp
}
\description{
Estimate ncp for Methyl Data
}
