% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputePCA.R
\name{methyl_imputePCA}
\alias{methyl_imputePCA}
\title{[imputePCA()] for methylation data}
\usage{
methyl_imputePCA(
  X,
  ncp = 2,
  scale = TRUE,
  method = c("Regularized", "EM"),
  coeff.ridge = 1,
  threshold = 1e-06,
  nb.init = 1,
  maxiter = 1000,
  ...
)
}
\value{
imputed matrix
}
\description{
[imputePCA()] for methylation data
}
