\name{missMDA-package}
\alias{missMDA-package}
\alias{missMDA}
\docType{package}
\title{Handling missing values with/in multivariate data analysis (principal component methods)
}
\description{
Imputation of incomplete continuous or categorical datasets; Missing values are imputed with a principal component analysis (PCA), a multiple correspondence analysis (MCA) model or a multiple factor analysis (MFA) model; Perform multiple imputation with and in PCA
}
\details{
\tabular{ll}{
Package: \tab missMDA\cr
Type: \tab Package\cr
Version: \tab 1.7.3\cr
Date: \tab 2014-11-26\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}

\author{
Francois Husson, Julie Josse

Maintainer: \email{husson@agrocampus-ouest.fr}
}

\references{
Josse, J. & Husson, F. (2012). Handling missing values in exploratory multivariate data analysis methods. Journal de la SFdS, 153(2), pp. 79-99.

Julie Josse, Francois Husson (2016). missMDA: A Package for Handling Missing Values in Multivariate Data Analysis. Journal of Statistical Software, 70(1), 1-31. doi:10.18637/jss.v070.i01
  
Some videos: \url{https://www.youtube.com/playlist?list=PLnZgp6epRBbQzxFnQrcxg09kRt-PA66T_}
}
\keyword{Factor analysis}

