% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_estim_ncpMCA.R
\name{modded_estim_ncpMCA}
\alias{modded_estim_ncpMCA}
\title{Estimate the Best `ncp` for `modded_imputeMCA` with cross-validation}
\usage{
modded_estim_ncpMCA(
  don,
  ncp.min = 0,
  ncp.max = 5,
  method.cv = c("Kfold", "loo"),
  nbsim = 100,
  pNA = 0.05,
  threshold = 1e-04,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{don:}{The input data.frame as a matrix or data frame.}

\item{ncp.min:}{The minimum number of dimensions to consider (default = 0).}

\item{ncp.max:}{The maximum number of dimensions to consider (default = 5).}

\item{method.cv:}{The cross-validation method to use, either "Kfold" or "loo" (default = "Kfold").}

\item{nbsim:}{The number of simulations to perform (default = 100).}

\item{pNA:}{The proportion of missing values to introduce in each simulation (default = 0.05).}

\item{threshold:}{The threshold value for convergence in the `imputeMCA` function (default = 1e-4).}

\item{verbose:}{Logical value indicating whether to display progress information (default = TRUE).}

\item{...:}{passed to `modded_imputeMCA`}
}
\value{
a list containing the estimated ncp and the corresponding mean criterion values over nbsim.
}
\description{
Estimates the number of dimensions to retain in Multiple Correspondence Analysis
}
\details{
This function estimates the number of dimensions to retain in MCA using the specified cross-validation method.
It calls either the `estim_ncpMCA_kfold` or `estim_ncpMCA_loo` function based on the method.cv argument.
This function first validates the input data.frame using the `validate_MCA` function.
This function then passes the necessary arguments to the selected cross-validation function and returns the result.
}
