% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeMCA.R
\name{modded_imputeMCA}
\alias{modded_imputeMCA}
\title{Modded imputeMCA by Refractoring}
\usage{
modded_imputeMCA(
  don,
  ncp = 2,
  coeff.ridge = 1,
  threshold = 1e-06,
  seed = NULL,
  maxiter = 1000
)
}
\arguments{
\item{don}{a data.frame with categorical variables containing missing values}

\item{ncp}{integer corresponding to the number of dimensions used to predict the missing entries}

\item{coeff.ridge}{1 by default to perform the regularized imputeMCA algorithm;
useful only if method="Regularized". Other regularization terms can be
implemented by setting the value to less than 1 in order to regularized less
(to get closer to the results of the EM method) or more than 1 to regularized
more (to get closer to the results of the proportion imputation)}

\item{threshold}{the threshold for assessing convergence}

\item{seed}{integer, by default seed = NULL implies that missing values are
initially imputed by the proportion of the category for the categorical
variables coded with indicator matrices of dummy variables. Other values
leads to a random initialization}

\item{maxiter}{integer, maximum number of iterations for the regularized iterative
MCA algorithm}
}
\value{
list of tab.disj and completeObs. See ?imputeMCA
}
\description{
Mod the imputeMCA function by omitting weights and sup variables
}
