% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_estim_ncpMCA.R
\name{prodna1}
\alias{prodna1}
\title{Insert NA into data.frame}
\usage{
prodna1(x, noNA)
}
\arguments{
\item{x}{The input data.frame}

\item{noNA}{The proportion of missing values to introduce, as a numeric value between 0 and 1.}
}
\value{
The input data.frame with randomly introduced missing values
}
\description{
Randomly introduces missing values (NA) into a given data.frame
}
\details{
This function creates a matrix of logical values (NAloc) with the same dimensions as the input data.frame.
For each column in the data.frame, it randomly selects a subset of row indices where the corresponding element is not an empty string.
The selected indices are used to assign TRUE values in the corresponding positions of the NAloc matrix.
The input data.frame is then updated by setting the elements corresponding to the TRUE values in NAloc as missing (NA).
The modified data.frame with missing values is returned.
}
